requirements.txt
amqp==5.3.1
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==43.0.3
cssselect2==0.7.0
Django==5.1.2
django-celery-beat==2.7.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
et_xmlfile==2.0.0
eventlet==0.38.0
greenlet==3.1.1
html5lib==1.1
idna==3.10
kombu==5.4.2
lxml==5.3.0
numpy==2.1.3
openpyxl==3.1.5
oscrypto==1.3.0
pandas==2.2.3
pillow==11.0.0
prompt_toolkit==3.0.48
pycparser==2.22
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
PyJWT==2.7.0
pypdf==5.1.0
python-bidi==0.6.3
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
redis==5.2.0
reportlab==4.2.5
requests==2.32.3
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.4.0
tzdata==2023.3
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
xhtml2pdf==0.2.16
